refacto: ContentDetailsLayout
