+++ packages/frontend/src/lib/ContentDetailsLayout.svelte
@@ -1,0 +1,41 @@
+<script lang="ts">
+export let detailsTitle: string;
+export let detailsLabel: string;
+let open: boolean = true;
+
+const toggle = () => {
+  open = !open;
+};
+</script>
+
+<div class="grid w-full overflow-y-auto lg:grid-cols-[1fr_auto] max-lg:grid-cols-[auto]">
+  <div class="p-5 inline-grid">
+    <slot name="content" />
+  </div>
+  <div class="inline-grid max-lg:order-first">
+    <div class="lg:my-5 max-lg:w-full max-lg:min-w-full" class:w-[375px]="{open}" class:min-w-[375px]="{open}">
+      <div
+        class:hidden="{!open}"
+        class:block="{open}"
+        class="h-fit lg:bg-charcoal-800 lg:rounded-l-md lg:mt-4 lg:py-4 max-lg:block"
+        aria-label="{`${detailsLabel} panel`}">
+        <div class="flex flex-col px-4 space-y-4 mx-auto">
+          <div class="w-full flex flex-row justify-between max-lg:hidden">
+            <span class="text-base">{detailsTitle}</span>
+            <button on:click="{toggle}" aria-label="{`hide ${detailsLabel}`}"
+              ><i class="fas fa-angle-right text-gray-900"></i></button>
+          </div>
+          <slot name="details" />
+        </div>
+      </div>
+      <div
+        class:hidden="{open}"
+        class:block="{!open}"
+        class="bg-charcoal-800 mt-4 p-4 rounded-md h-fit max-lg:hidden"
+        aria-label="{`toggle ${detailsLabel}`}">
+        <button on:click="{toggle}" aria-label="{`show ${detailsLabel}`}"
+          ><i class="fas fa-angle-left text-gray-900"></i></button>
+      </div>
+    </div>
+  </div>
+</div>
